Move message class into extension/background folder
