+++ packages/sanes-chrome-extension/src/index.tsx
@@ -1,13 +1,13 @@
 import { ConnectedRouter } from 'connected-react-router';
-import MedulasThemeProvider from 'medulas-react-components/lib/theme/MedulasThemeProvider';
 import { ToastProvider } from 'medulas-react-components/lib/context/ToastProvider';
+import MedulasThemeProvider from 'medulas-react-components/lib/theme/MedulasThemeProvider';
 import * as React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import { PersonaProvider } from './context/PersonaProvider';
-import { sendGetPersonaMessage, GetPersonaResponse } from './extension/messages';
+import { GetPersonaResponse, sendGetPersonaMessage } from './extension/background/messages';
 import Route from './routes';
-import { WELCOME_ROUTE, ACCOUNT_STATUS_ROUTE } from './routes/paths';
+import { ACCOUNT_STATUS_ROUTE, WELCOME_ROUTE } from './routes/paths';
 import { makeStore } from './store';
 import { history } from './store/reducers';
 import { globalStyles } from './theme/globalStyles';
